chore: move profiles controller
